Use explicit argument types
- beginning conversion of argument types to explicit types
- implements conversion operators and extended vector types with ctor and components
- add all explicit argument types to dictionary
- finished conversion of KVNucleus to explicit argument types
- This version only works with ROOT6:

Error caused by overloaded operator<< for KVArgType 'TString' variables.
The error comes at compilation of the dictionary for KVMultiDetbase.
Compiler complains there is no 'ostream& operator<<(ostream&,const char*)'
even though of course this exists.
Adding '--std=c++11' does not help, it looks like this is a problem with
the dictionary generated by rootcint which does not occur when rootcling
is used (ROOT6)

+++ KVMultiDet/events/KVGenPhaseSpace.cpp
@@ -173,7 +173,7 @@ Bool_t KVGenPhaseSpace::CheckBreakUpChannel()
    Int_t atot = (Int_t)fEvent->GetSum("GetA");
    if (ztot != fCompound.GetZ() || atot != fCompound.GetA()) {
       Warning("CheckBreakUpChannel", "Compound has (Z,A)=(%d,%d), break-up channel has (Z,A)=(%d,%d)",
-              fCompound.GetZ(), fCompound.GetA(), ztot, atot);
+              (int)fCompound.GetZ(), (int)fCompound.GetA(), ztot, atot);
       return kFALSE;
    }
